requirements.txt
opencv-python
numpy
pyyaml
torch
ultralytics